stake-odds/requirements.txt
pillow>=10.0.0
pytesseract>=0.3.10
mss>=9.0.0
treys>=0.1.8
rich>=13.0.0
numpy>=1.24.0
